requirements.txt
# Core Deep Learning
tensorflow>=2.16.0
keras>=3.0.0

# Data Processing
numpy==1.24.3
pandas==2.0.3
opencv-python==4.8.0.74
Pillow==10.0.0
scikit-learn==1.3.0
scikit-image==0.21.0
imbalanced-learn==0.11.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Web Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0

# Database
pymysql==1.1.0
cryptography==41.0.3

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
tqdm==4.66.1
joblib==1.3.2

# Grad-CAM and Explainability
tf-keras-vis==0.8.5

# Image Augmentation
albumentations==1.3.1

# Jupyter
jupyter==1.0.0
ipykernel==6.25.1
notebook==7.0.2

# Deployment
gunicorn==21.2.0
waitress==2.1.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0
